+++ packages/nimble-blazor/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-blazor
 
-This log was last generated on Mon, 01 Aug 2022 17:08:37 GMT and should not be manually modified.
+This log was last generated on Tue, 09 Aug 2022 19:30:28 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 5.3.5
+
+Tue, 09 Aug 2022 19:30:28 GMT
+
+### Patches
+
+- Adding 'Placeholder' APIs to Combobox. ([ni/nimble@2b7f674](https://github.com/ni/nimble/commit/2b7f6744f4c9f1504b3852aa059f2b1f1676222c))
 
 ## 5.3.0